01-mathquiz.py
This code section configures the Maths Quiz app's audio and graphics.  In order to handle errors gracefully, it sets up Pygame to play background music and sound effects for both right and wrong answers.  Additionally, it uses PIL to load and resize background images for Easy, Moderate, and Advanced modes. These images are then stored in a dictionary for dynamic switching while playing.